Fix Protect Session Cookies
In file: csrf.go, a cookie is being constructed keeping the {HttpOnly} or the {Secure} flags set to false. The {HttpOnly} flag prevents the cookie from being read by any client-side scripts and the {Secure} flag ensures that the cookie will be transmitted via a secure connection. iCR fixed the issue by creating a cookie with {HttpOnly} and {Secure} flags set to true.